fix: Move env fetcher script out of load-config-from-aws
