+++ .github/workflows/build-and-push-docker-image-to-registry.yml
@@ -2,7 +2,7 @@ name: Build and Push Docker Image
 
 on:
   push:
-    branches: ['main']
+    branches: ['main', 'fix/solve-env-issues']
 
 env:
   REGISTRY: ghcr.io
